dont cast for std:: one

+++ mini_chromium/mini_chromium/base/atomicops_internals_portable.h
@@ -56,7 +56,7 @@ inline void MemoryBarrier() {
   // not defined, leading to the linker complaining about undefined references.
   __atomic_thread_fence(static_cast<int>(std::memory_order_seq_cst));
 #else
-  std::atomic_thread_fence(static_cast<int>(std::memory_order_seq_cst));
+  std::atomic_thread_fence(std::memory_order_seq_cst);
 #endif
 }
